Use double in decimal_to_fraction_cont to avoid integer overflow

+++ src/decimal_to_fraction.cpp
@@ -2,38 +2,38 @@
 using namespace Rcpp;
 
 IntegerVector decimal_to_fraction_cont(const double x, const long max_denom) {
-  long      n0 = 0;
-  long      n1 = 1;
-  long long n2 = floor(x);
-  long      d0 = 0;
-  long      d1 = 0;
-  long long d2 = 1;
-  long long f  = n2;
-  double    z  = x - n2;
+  double n0 = 0;
+  double n1 = 1;
+  double n2 = floor(x);
+  double d0 = 0;
+  double d1 = 0;
+  double d2 = 1;
+  double f  = n2;
+  double z  = x - n2;
 
   while (d2 <= max_denom) {
     z  = 1 / z;
-    f  = (long long) floor(z);
-    z  = z - (double) f;
+    f  = floor(z);
+    z  = z - f;
     n0 = n1;
-    n1 = (long) n2;
-    n2 = f * (long long) n1 + (long long) n0;
+    n1 = n2;
+    n2 = f * n1 + n0;
     d0 = d1;
-    d1 = (long) d2;
-    d2 = f * (long long) d1 + (long long) d0;
-    if (fabs(x - (double) n1 / (double) d1) < DBL_EPSILON) {break;}
+    d1 = d2;
+    d2 = f * d1 + d0;
+    if (fabs(x - n1 / d1) < DBL_EPSILON) {break;}
     if (f == 0) {break;}
   }
 
   if (n1 == LONG_MIN) {
-    n1 = (long) ceil(x);
+    n1 = ceil(x);
     d1 = 1;
   } else if (d1 == LONG_MIN) {
-    n1 = (long) floor(x);
+    n1 = floor(x);
     d1 = 1;
   }
 
-  return IntegerVector::create(n1, d1);
+  return IntegerVector::create((int) n1, (int) d1);
 }
 
 IntegerVector decimal_to_fraction_base_10(const double x, const long max_denom)
